requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-Limiter==3.5.0
WTForms==3.1.0
psycopg2-binary==2.9.7
redis==5.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
APScheduler==3.10.4
bcrypt==4.0.1
email-validator==2.1.0